Remove segment from list and add undo/redo for segments (#6944)
* implement remove-segment-from-list and add undo/redo when renaming or removing a segment

* rename most occurrences of cellId to segmentId

* don't make icons italic

* fix missing reset of prevSegments and also enforce empty currentVolumeUndoBuckets constraint

* reorder if-conditions to avoid unnecessary check of actions nullity

* move undo saga related code into own module

* remove unused var

* add more comments

* fix undo/redo for meshes

* add comment

* change undo mechanism so that visible changes to the segment list that were caused by direct actions of the user create new undo states

* update changelog
