+++ frontend/javascripts/oxalis/model/reducers/annotation_reducer.ts
@@ -249,8 +249,8 @@ function AnnotationReducer(state: OxalisState, action: Action): OxalisState {
     }
 
     case "REMOVE_ISOSURFACE": {
-      const { layerName, cellId } = action;
-      const { [cellId]: _, ...remainingIsosurfaces } =
+      const { layerName, segmentId } = action;
+      const { [segmentId]: _, ...remainingIsosurfaces } =
         state.localSegmentationData[layerName].isosurfaces;
       return updateKey2(state, "localSegmentationData", layerName, {
         isosurfaces: remainingIsosurfaces,
@@ -258,9 +258,9 @@ function AnnotationReducer(state: OxalisState, action: Action): OxalisState {
     }
 
     case "ADD_AD_HOC_ISOSURFACE": {
-      const { layerName, cellId, seedPosition, mappingName, mappingType } = action;
+      const { layerName, segmentId, seedPosition, mappingName, mappingType } = action;
       const isosurfaceInfo: IsosurfaceInformation = {
-        segmentId: cellId,
+        segmentId: segmentId,
         seedPosition,
         isLoading: false,
         isVisible: true,
@@ -273,15 +273,15 @@ function AnnotationReducer(state: OxalisState, action: Action): OxalisState {
         "localSegmentationData",
         layerName,
         "isosurfaces",
-        cellId,
+        segmentId,
         isosurfaceInfo,
       );
     }
 
     case "ADD_PRECOMPUTED_ISOSURFACE": {
-      const { layerName, cellId, seedPosition, meshFileName } = action;
+      const { layerName, segmentId, seedPosition, meshFileName } = action;
       const isosurfaceInfo: IsosurfaceInformation = {
-        segmentId: cellId,
+        segmentId: segmentId,
         seedPosition,
         isLoading: false,
         isVisible: true,
@@ -293,13 +293,13 @@ function AnnotationReducer(state: OxalisState, action: Action): OxalisState {
         "localSegmentationData",
         layerName,
         "isosurfaces",
-        cellId,
+        segmentId,
         isosurfaceInfo,
       );
     }
 
     case "STARTED_LOADING_ISOSURFACE": {
-      const { layerName, cellId } = action;
+      const { layerName, segmentId } = action;
       const isosurfaceInfo: Partial<IsosurfaceInformation> = {
         isLoading: true,
       };
@@ -308,13 +308,13 @@ function AnnotationReducer(state: OxalisState, action: Action): OxalisState {
         "localSegmentationData",
         layerName,
         "isosurfaces",
-        cellId,
+        segmentId,
         isosurfaceInfo,
       );
     }
 
     case "FINISHED_LOADING_ISOSURFACE": {
-      const { layerName, cellId } = action;
+      const { layerName, segmentId } = action;
       const isosurfaceInfo: Partial<IsosurfaceInformation> = {
         isLoading: false,
       };
@@ -323,7 +323,7 @@ function AnnotationReducer(state: OxalisState, action: Action): OxalisState {
         "localSegmentationData",
         layerName,
         "isosurfaces",
-        cellId,
+        segmentId,
         isosurfaceInfo,
       );
     }
