+++ frontend/javascripts/oxalis/model/sagas/root_saga.ts
@@ -3,6 +3,7 @@ import { all, call, cancel, fork, take } from "typed-redux-saga";
 import { alert } from "libs/window";
 import VolumetracingSagas from "oxalis/model/sagas/volumetracing_saga";
 import SaveSagas, { toggleErrorHighlighting } from "oxalis/model/sagas/save_saga";
+import UndoSaga from "oxalis/model/sagas/undo_saga";
 import AnnotationSagas from "oxalis/model/sagas/annotation_saga";
 import { watchDataRelevantChanges } from "oxalis/model/sagas/prefetch_saga";
 import SkeletontracingSagas from "oxalis/model/sagas/skeletontracing_saga";
@@ -51,6 +52,7 @@ function* restartableSaga(): Saga<void> {
       call(ProofreadSaga),
       ...AnnotationSagas.map((saga) => call(saga)),
       ...SaveSagas.map((saga) => call(saga)),
+      call(UndoSaga),
       ...VolumetracingSagas.map((saga) => call(saga)),
       call(watchZ1Downsampling),
     ]);
