+++ frontend/javascripts/oxalis/view/right-border-tabs/segments_tab/segment_list_item.tsx
@@ -154,7 +154,13 @@ type Props = {
   handleSegmentDropdownMenuVisibility: (arg0: number, arg1: boolean) => void;
   activeDropdownSegmentId: number | null | undefined;
   allowUpdate: boolean;
-  updateSegment: (arg0: number, arg1: Partial<Segment>, arg2: string) => void;
+  updateSegment: (
+    arg0: number,
+    arg1: Partial<Segment>,
+    arg2: string,
+    createsNewUndoState: boolean,
+  ) => void;
+  removeSegment: (arg0: number, arg2: string) => void;
   onSelectSegment: (arg0: Segment) => void;
   visibleSegmentationLayer: APISegmentationLayer | null | undefined;
   loadAdHocMesh: (arg0: number, arg1: Vector3) => void;
@@ -191,16 +197,13 @@ function _MeshInfoItem(props: {
   };
 
   const { segment, isSelectedInList, isHovered, isosurface } = props;
-  const deemphasizedStyle = {
-    fontStyle: "italic",
-    color: "#989898",
-  };
 
   if (!isosurface) {
     if (isSelectedInList) {
       return (
         <div
-          style={{ ...deemphasizedStyle, marginLeft: 8 }}
+          className="deemphasized italic"
+          style={{ marginLeft: 8 }}
           onContextMenu={(evt) => {
             evt.preventDefault();
             props.handleSegmentDropdownMenuVisibility(segment.id, true);
@@ -215,7 +218,7 @@ function _MeshInfoItem(props: {
   }
 
   const { seedPosition, isLoading, isPrecomputed, isVisible } = isosurface;
-  const textStyle = isVisible ? {} : deemphasizedStyle;
+  const className = isVisible ? "" : "deemphasized italic";
   const downloadButton = (
     <Tooltip title="Download Mesh">
       <VerticalAlignBottomOutlined
@@ -281,10 +284,11 @@ function _MeshInfoItem(props: {
         >
           {toggleVisibilityCheckbox}
           <span
+            className={className}
             onClick={() => {
               props.setPosition(seedPosition);
             }}
-            style={{ ...textStyle, marginLeft: 8 }}
+            style={{ marginLeft: 8 }}
           >
             {isPrecomputed ? "Mesh (precomputed)" : "Mesh (ad-hoc computed)"}
           </span>
@@ -320,6 +324,7 @@ function _SegmentListItem({
   activeDropdownSegmentId,
   allowUpdate,
   updateSegment,
+  removeSegment,
   onSelectSegment,
   visibleSegmentationLayer,
   loadAdHocMesh,
@@ -363,20 +368,20 @@ function _SegmentListItem({
         andCloseContextMenu,
       ),
       getMakeSegmentActiveMenuItem(segment, setActiveCell, activeCellId, andCloseContextMenu),
-      /*
-       * Disable the change-color menu if the segment was mapped to another segment, because
-       * changing the color wouldn't do anything as long as the mapping is still active.
-       * This is because the id (A) is mapped to another one (B). So, the user would need
-       * to change the color of B to see the effect for A.
-       */
       {
         key: "changeSegmentColor",
+        /*
+         * Disable the change-color menu if the segment was mapped to another segment, because
+         * changing the color wouldn't do anything as long as the mapping is still active.
+         * This is because the id (A) is mapped to another one (B). So, the user would need
+         * to change the color of B to see the effect for A.
+         */
         disabled: isEditingDisabled || segment.id !== mappedId,
         label: (
           <ChangeColorMenuItemContent
             isDisabled={isEditingDisabled}
             title="Change Segment Color"
-            onSetColor={(color) => {
+            onSetColor={(color, createsNewUndoState) => {
               if (visibleSegmentationLayer == null) {
                 return;
               }
@@ -386,6 +391,7 @@ function _SegmentListItem({
                   color,
                 },
                 visibleSegmentationLayer.name,
+                createsNewUndoState,
               );
             }}
             rgb={Utils.take3(segmentColorRGBA)}
@@ -406,9 +412,22 @@ function _SegmentListItem({
               color: null,
             },
             visibleSegmentationLayer.name,
+            true,
           );
         },
         label: "Reset Segment Color",
+      },
+      {
+        key: "removeSegmentFromList",
+        disabled: isEditingDisabled,
+        onClick: () => {
+          if (isEditingDisabled || visibleSegmentationLayer == null) {
+            return;
+          }
+          removeSegment(segment.id, visibleSegmentationLayer.name);
+          andCloseContextMenu();
+        },
+        label: "Remove Segment From List",
       },
     ],
   });
@@ -417,7 +436,7 @@ function _SegmentListItem({
     if (isJSONMappingEnabled && segment.id !== mappedId)
       return (
         <Tooltip title="Segment ID (Unmapped ID → Mapped ID)">
-          <span className="deemphasized-segment-name">
+          <span className="deemphasized italic">
             {segment.id} → {mappedId}
           </span>
         </Tooltip>
@@ -425,7 +444,7 @@ function _SegmentListItem({
     // Only if segment.name is truthy, render additional info.
     return segment.name ? (
       <Tooltip title="Segment ID">
-        <span className="deemphasized-segment-name">{segment.id}</span>
+        <span className="deemphasized italic">{segment.id}</span>
       </Tooltip>
     ) : null;
   }
@@ -475,6 +494,7 @@ function _SegmentListItem({
                       name,
                     },
                     visibleSegmentationLayer.name,
+                    true,
                   );
                 }
               }}
@@ -491,7 +511,7 @@ function _SegmentListItem({
             {segment.id === centeredSegmentId ? (
               <Tooltip title="This segment is currently centered in the data viewports.">
                 <i
-                  className="fas fa-crosshairs deemphasized-segment-name"
+                  className="fas fa-crosshairs deemphasized"
                   style={{
                     marginLeft: 4,
                   }}
@@ -501,7 +521,7 @@ function _SegmentListItem({
             {segment.id === activeCellId ? (
               <Tooltip title="The currently active segment id belongs to this segment.">
                 <i
-                  className="fas fa-paint-brush deemphasized-segment-name"
+                  className="fas fa-paint-brush deemphasized"
                   style={{
                     marginLeft: 4,
                   }}
