+++ frontend/stylesheets/_utils.less
@@ -126,3 +126,11 @@ td.nowrap * {
 .text-center {
   text-align: center;
 }
+
+.deemphasized {
+  color: #989898;
+}
+
+.italic {
+  font-style: italic;
+}
