+++ frontend/stylesheets/trace_view/_right_menu.less
@@ -51,11 +51,6 @@
   }
 }
 
-.deemphasized-segment-name {
-  font-style: italic;
-  color: #989898;
-}
-
 #commentList {
   .comment,
   .comment-tree {
